+++ mobile/lib/presentation/screens/paiement/receive_screen.dart
@@ -54,7 +54,7 @@ class _ReceiveScreenState extends State<ReceiveScreen> {
       return;
     }
 
-    final montant = int.tryParse(_montantController.text) ?? 0;
+    final montant = double.tryParse(_montantController.text) ?? 0;
 
     setState(() {
       _paymentRequest = PaymentRequest(
